Moves the new methods to Imaging.Helpers.ImageMaths and auto-cleanup (reordering of methods) of the ImageMaths class
